requirements.txt
aiohttp
certifi
discord-py-interactions==5.13.2
ddddocr==1.5.6